--- src/pages/config/settings/components/scrape-run-viewer/viewer/globals.d.ts
@@ -1,13 +1,0 @@
-// Build-time constants injected by vite via `define`. See vite.config.ts.
-declare const __UI_BUILD_COMMIT__: string;
-declare const __UI_BUILD_DATE__: string;
-
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      "iconify-icon": any;
-    }
-  }
-}
-
-export {};
