+++ src/pages/config/settings/components/viewer/components/DetailPanel.tsx
