+++ src/pages/config/settings/components/viewer/components/HARPanel.tsx
