+++ src/pages/config/settings/components/viewer/components/SnapshotPanel.tsx
