+++ src/pages/config/settings/components/viewer/components/SplitPane.tsx
