+++ src/pages/config/settings/components/viewer/globals.d.ts
@@ -1,0 +1,9 @@
+declare global {
+  namespace JSX {
+    interface IntrinsicElements {
+      "iconify-icon": any;
+    }
+  }
+}
+
+export {};
