+++ src/pages/config/settings/components/viewer/types.ts
